+++ execution_engine/src/core/runtime/mod.rs
@@ -1045,6 +1045,14 @@ where
             }
         };
 
+        if let wasmi::Error::Trap(Trap::Code(TrapCode::Unreachable)) = &error {
+            if let Some(stack_height) = instance.globals().last() {
+                eprintln!("current stack height = {}", stack_height.get())
+            } else {
+                eprintln!("current stack height unknown");
+            }
+        }
+
         if let Some(host_error) = error.as_host_error() {
             // If the "error" was in fact a trap caused by calling `ret` then
             // this is normal operation and we should return the value captured
@@ -1380,6 +1388,14 @@ where
                 return Ok(runtime.take_host_buffer().unwrap_or(CLValue::from_t(())?));
             }
         };
+
+        if let wasmi::Error::Trap(Trap::Code(TrapCode::Unreachable)) = &error {
+            if let Some(stack_height) = instance.globals().last() {
+                eprintln!("current stack height = {}", stack_height.get())
+            } else {
+                eprintln!("current stack height unknown");
+            }
+        }
 
         if let Some(host_error) = error.as_host_error() {
             // If the "error" was in fact a trap caused by calling `ret` then this is normal